fix: (bridge) Some JS features unsupported on older browsers
The bridge script was not working on Android 5 and 6 because the default install of Chrome is a very outdated version. This means our bridge script needs to use features which target ES5. This commit replaces arrow functions (=>) with normal functions and replaces 'Array.find' with 'Array.filter'. Also moved the 'module.exports' to an 'if' statement to avoid errors in the browser.