docs: Fix name appearing in generated typedocs
The name was showing up in typedocs as 'js-miniapp-bridge' and also it was showing the version number for js-miniapp-bridge instead of the version for js-miniapp-sdk. This is possibly a bug in typedocs where it's pulling this from an external module rather than the main module. The only fix I could find is to manually set the 'name' property and remove the version.

Also moved the typedoc options into a typedoc.json file