fix: Add date to versions to fix order of the list (#147)
We had an issue where the versions are being sorted by name, but Jekyll cannot recognize it as semantic versions so the sorting is incorrect for versions greater than X.10. Jekyll will sort by date if the a date exists for each item, so we added the date to each item to fix the sorting.