### [MINI-5752] Universal Bridge JS Sample app changes to handle response form Host app for Interface 1.
Issue:  The returned response from Host app for sendJsonToHostapp(info) api was showing alert, this impacts the interface-1 feature in iOS case.

Cause: response of sendJsonToHostapp() was handled as window.alert().

Fix: Changed to handle the response form host app sdk sendJsonToHostapp() api to be show as text label in the js-sample miniapp.

**Link:**
[MINI-5752](https://jira.rakuten-it.com/jira/browse/MINI-5752)